fix: avoid dialog mock startup race
