+++ packages/renderer-worker/test/SendMessagePortToExtensionHostWorker.test.ts
@@ -11,21 +11,9 @@ jest.unstable_mockModule('../src/parts/SharedProcess/SharedProcess.js', () => {
   }
 })
 
-jest.unstable_mockModule('../src/parts/DialogWorker/DialogWorker.js', () => {
+jest.unstable_mockModule('../src/parts/HandleDialogWorkerMessagePort/HandleDialogWorkerMessagePort.ts', () => {
   return {
-    invokeAndTransfer: jest.fn(),
-  }
-})
-
-jest.unstable_mockModule('../src/parts/ConfirmPrompt/ConfirmPrompt.js', () => {
-  return {
-    isMocked: jest.fn(() => false),
-  }
-})
-
-jest.unstable_mockModule('../src/parts/HandleMockDialogWorkerMessagePort/HandleMockDialogWorkerMessagePort.ts', () => {
-  return {
-    handleMockDialogWorkerMessagePort: jest.fn(),
+    handleDialogWorkerMessagePort: jest.fn(),
   }
 })
 
@@ -65,11 +53,9 @@ jest.unstable_mockModule('../src/parts/WorkspaceConnection/WorkspaceConnection.j
   }
 })
 
-const ConfirmPrompt = await import('../src/parts/ConfirmPrompt/ConfirmPrompt.js')
-const DialogWorker = await import('../src/parts/DialogWorker/DialogWorker.js')
 const ExtensionManagementWorker = await import('../src/parts/ExtensionManagementWorker/ExtensionManagementWorker.js')
 const ExplorerViewWorker = await import('../src/parts/ExplorerViewWorker/ExplorerViewWorker.js')
-const HandleMockDialogWorkerMessagePort = await import('../src/parts/HandleMockDialogWorkerMessagePort/HandleMockDialogWorkerMessagePort.ts')
+const HandleDialogWorkerMessagePort = await import('../src/parts/HandleDialogWorkerMessagePort/HandleDialogWorkerMessagePort.ts')
 const MainAreaWorker = await import('../src/parts/MainAreaWorker/MainAreaWorker.js')
 const SecretsViewWorker = await import('../src/parts/SecretsViewWorker/SecretsViewWorker.ts')
 const SettingsWorker = await import('../src/parts/SettingsWorker/SettingsWorker.js')
@@ -113,23 +99,12 @@ test('sendMessagePortToTerminalProcess uses the workspace connection', async () 
   expect(SharedProcess.invokeAndTransfer).not.toHaveBeenCalled()
 })
 
-test('sendMessagePortToDialogWorker', async () => {
-  const port = {}
-
-  await SendMessagePortToExtensionHostWorker.sendMessagePortToDialogWorker(port, 'HandleMessagePort.handleMessagePort')
-
-  expect(DialogWorker.invokeAndTransfer).toHaveBeenCalledTimes(1)
-  expect(DialogWorker.invokeAndTransfer).toHaveBeenCalledWith('HandleMessagePort.handleMessagePort', port)
-})
-
-test('sendMessagePortToDialogWorker preserves confirm prompt mocks', async () => {
+test('sendMessagePortToDialogWorker routes prompts through the renderer worker', async () => {
   const { port1: port, port2 } = new MessageChannel()
-  jest.mocked(ConfirmPrompt.isMocked).mockReturnValue(true)
 
   await SendMessagePortToExtensionHostWorker.sendMessagePortToDialogWorker(port, 'HandleMessagePort.handleMessagePort')
 
-  expect(HandleMockDialogWorkerMessagePort.handleMockDialogWorkerMessagePort).toHaveBeenCalledWith(port)
-  expect(DialogWorker.invokeAndTransfer).not.toHaveBeenCalled()
+  expect(HandleDialogWorkerMessagePort.handleDialogWorkerMessagePort).toHaveBeenCalledWith(port)
   port.close()
   port2.close()
 })
